+++ js/main.js
@@ -1,9 +1,8 @@
 import { getPhotos } from './data.js';
 import { renderPhotos } from './photos.js';
-import './full-photo.js';
+import { addPicturesListener } from './gallery.js';
 
 const photos = getPhotos();
 
 renderPhotos(photos);
-
-export {photos};
+addPicturesListener(photos);
